ecom_airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Core Airflow and Providers
# astronomer-cosmos
# apache-airflow-providers-postgres
# apache-airflow-providers-amazon
# faker==19.3.0
# boto3==1.35.77
# psycopg2-binary==2.9.10
# python-dotenv==1.0.1
# SQLAlchemy
# pyarrow==18.1.0

# python-dotenv
# apache-airflow-providers-dbt-cloud>=3.3.0
# astronomer-cosmos[duckdb]>=1.0.3
# dbt-core>=1.7.1
# dbt-duckdb>=1.7.1

astronomer-cosmos
apache-airflow-providers-postgres
apache-airflow-providers-amazon
airflow-provider-duckdb==0.2.0
duckdb==1.1.3
faker==19.3.0
boto3==1.35.77
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy
pyarrow==18.1.0
